backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# CORS
django-cors-headers>=4.3.0

# Environment
python-dotenv>=1.0.0
django-environ>=0.11.0

# File Processing
pdfplumber>=0.10.0
python-docx>=1.1.0

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0

# Async Tasks
celery[redis]>=5.3.0
redis>=5.0.0

# Storage (for production)
django-storages>=1.14.0
boto3>=1.28.0

# Utilities
Pillow>=10.0.0

